Python Agents/pyproject.toml
[project]
name = "learning_python_agents"
version = "0.1.0"
description = "Python Agents"
authors = [
    { name = "Alessandro Piccione" }
]
requires-python = ">=3.12"
dependencies = [
    "agent-framework",
    "asyncio>=4.0.0",
    "opentelemetry-semantic-conventions-ai==0.4.13",  # Pinned version for error in agent-framework"
    "python-dotenv>=1.2.1",
    "rich>=14.3.3",
]


#[build-system]
#requires = ["hatchling"]
#build-backend = "hatchling.build"

#[tool.hatch.build.targets.wheel]
## if the code is in the root folder us "."
#packages = ["src"]

[tool.uv]
#prerelease = "allow"

